--- application/Core/views/scripts/error/error.phtml
@@ -1,24 +1,0 @@
-<center>
-	<h2>An error occurred</h2>
-</center>
-
-
-<h1><tal:block content="message" /></h1>
-
-<tal:block condition="exists: exception">
-
-	<h3>Exception information:</h3>
-	<p>
-		<b>Message:</b>
-		<tal:block content="php: exception.getMessage()" />
-	</p>
-
-
-	<h3>Stack trace:</h3>
-	<pre tal:content="php: exception.getTraceAsString()" />
-
-
-	<h3>Request Parameters:</h3>
-	<pre tal:content="php: var_export( request.getParams(), true )" />
-
-</tal:block>
