+++ application/Module/WebApp/Controller/GroupController.php
@@ -86,7 +86,7 @@ class WebApp_GroupController extends \WebApp\Controller\BaseController
 	}
 
 	public function newcampAction(){
-		$form = new \Core\Form\Camp();
+		$form = new \WebApp\Form\Camp();
 		
 		$form->setDefaults($this->getRequest()->getParams());
 
@@ -95,7 +95,7 @@ class WebApp_GroupController extends \WebApp\Controller\BaseController
 
 	public function createcampAction()
 	{
-		$form = new \Core\Form\Camp();
+		$form = new \WebApp\Form\Camp();
 		$params = $this->getRequest()->getParams();
 		
 		if(!$form->isValid($params))
@@ -122,7 +122,7 @@ class WebApp_GroupController extends \WebApp\Controller\BaseController
 	
 	public function searchuserAction()
 	{
-		$search = new \Core\Form\Search();
+		$search = new \WebApp\Form\Search();
 		$query = "";
 		$users = array();
